V17.35
Updated the low stock report page to include products that have never
been through the goods in system. this should fix the issue where by new
products added to the database not showing up on the low stock report
due to the date comparison in place.

updated the myApp.js for the shredmaster calculator